rationalise locations of readwrite code
