+++ andy_elf/drivers/ata_pio/interrupt.c
@@ -6,11 +6,9 @@
 #include <panic.h>
 #include <memops.h>
 #include "ata_pio.h"
+#include "ata_readwrite.h"
 
 extern ATADriverState *master_driver_state;
-
-void ata_complete_read_lowerhalf(SchedulerTask *t);
-void ata_complete_write_lowerhalf(SchedulerTask *t);
 
 /*
 This is called from the (assembly language) IPrimaryATA / ISecondaryATA
@@ -57,86 +55,4 @@ void ata_service_interrupt(uint8_t bus_nr)
 void ata_dump_errors(uint8_t err)
 {
   kprintf("ERROR ATA error code %d\r\n", (uint16_t)err);
-}
-
-/*
-This is called by the scheduler from the kernel context
-*/
-void ata_complete_read_lowerhalf(SchedulerTask *t)
-{
-  ATAPendingOperation *op = (ATAPendingOperation *)t->data;
-
-  kprintf("DEBUG in disk read lower-half. Requested paging dir is 0x%x\r\n", op->paging_directory);
-
-  vaddr old_pd = switch_paging_directory_if_required((vaddr)op->paging_directory);
-
-  //need to make sure interrupts are disabled, otherwise we trigger the next data packet
-  //before we stored the last word of this one, meaning that we miss data.
-  cli();
-  //each sector is 512 bytes (or 256 words)
-  uint16_t *buf = (uint16_t *)op->buffer;
-  for(register size_t i=op->buffer_loc; i<op->buffer_loc+256;i++) {
-    if(i==op->buffer_loc+255) inb(ATA_STATUS(op->base_addr)); //read status byte to reset interrupt flag. We don't actually care about the values right now.
-    buf[i] = inw(ATA_DATA_REG(op->base_addr));
-  }
-
-  op->buffer_loc += 256;
-  ++op->sectors_read;
-  if(op->sectors_read>=op->sector_count) {
-    //reset the "pending operation" block for the next operation
-    //memset((void *)op, 0, sizeof(ATAPendingOperation));
-    op->type = ATA_OP_NONE;
-    //the operation is now completed
-    op->completed_func(ATA_STATUS_OK, op->buffer);
-  }
-
-  if(old_pd!=NULL) switch_paging_directory_if_required(old_pd);
-
-  sti();
-
-}
-
-void ata_continue_write(ATAPendingOperation *op)
-{
-  vaddr old_pd = switch_paging_directory_if_required((vaddr)op->paging_directory);
-
-  //need to make sure interrupts are disabled, otherwise we trigger the next data packet
-  //before we stored the last word of this one, meaning that we miss data.
-  cli();
-
-  kprintf("DEBUG sectors written %d sectors total to write %d\r\n", op->sectors_read, op->sector_count);
-
-  //each sector is 512 bytes (or 256 words)
-  uint16_t *buf = (uint16_t *)op->buffer;
-  for(register size_t i=op->buffer_loc; i<op->buffer_loc+256;i++) {
-    if(i==op->buffer_loc+255) inb(ATA_STATUS(op->base_addr)); //read status byte to reset interrupt flag. We don't actually care about the values right now.
-    outw(ATA_DATA_REG(op->base_addr), buf[i]);
-    io_wait();  //in ioports.asm
-  }
-
-  outb(ATA_COMMAND(op->base_addr), ATA_CMD_CACHE_FLUSH);  //make sure to flush the cache once the write has completed
-
-  op->buffer_loc += 256;
-  ++op->sectors_read;
-
-
-  if(old_pd!=NULL) switch_paging_directory_if_required(old_pd);
-
-  sti();
-}
-
-void ata_complete_write_lowerhalf(SchedulerTask *t)
-{
-  ATAPendingOperation *op = (ATAPendingOperation *)t->data;
-
-  kprintf("DEBUG in disk write lower-half. Requested paging dir is 0x%x\r\n", op->paging_directory);
-
-  if(op->sectors_read>=op->sector_count) {
-    //the operation is now completed
-    op->completed_func(ATA_STATUS_OK, op->buffer);
-    //reset the "pending operation" block for the next operation
-    op->type = ATA_OP_NONE;
-  } else {
-      ata_continue_write(op);
-  }
 }
