+++ andy_elf/drivers/ata_pio/readwrite.c
@@ -7,6 +7,7 @@
 #include <errors.h>
 #include <panic.h>
 #include "ata_pio.h"
+#include "ata_readwrite.h"
 
 extern ATADriverState *master_driver_state;
 
@@ -104,9 +105,6 @@ int8_t ata_pio_start_read(uint8_t drive_nr, uint64_t lba_address, uint8_t sector
   outb(ATA_COMMAND(base_addr), ATA_CMD_READ_SECTORS);
 }
 
-//this is currently defined in interrupt.c, it's just the write-words-to-output-port
-void ata_continue_write(ATAPendingOperation *op);
-
 int8_t ata_pio_start_write(uint8_t drive_nr, uint64_t lba_address, uint8_t sector_count, void *buffer, void (*callback)(uint8_t status, void *buffer))
 {
   uint16_t base_addr;
@@ -152,4 +150,109 @@ int8_t ata_pio_start_write(uint8_t drive_nr, uint64_t lba_address, uint8_t secto
   outb(ATA_COMMAND(base_addr), ATA_CMD_WRITE_SECTORS);
 
   ata_continue_write(op); //write the first sector. Second one will get set by the interrupt.
+}
+
+/*
+This is a "lower half", i.e. a routine that is run by the scheduler from the root
+kernel event loop. It's requested by the interrupt handler and does the actual work
+of reading bytes from the disk and transferring them into memory.
+Arguments: t - pointer to the SchedulerTask that was used to queue the event. The
+      "data" field can be assumed to be a pointer to ATAPendingOperation which
+      gives details of the event in progress
+Returns: Nothing
+*/
+void ata_complete_read_lowerhalf(SchedulerTask *t)
+{
+  ATAPendingOperation *op = (ATAPendingOperation *)t->data;
+
+  kprintf("DEBUG in disk read lower-half. Requested paging dir is 0x%x\r\n", op->paging_directory);
+
+  vaddr old_pd = switch_paging_directory_if_required((vaddr)op->paging_directory);
+
+  //need to make sure interrupts are disabled, otherwise we trigger the next data packet
+  //before we stored the last word of this one, meaning that we miss data.
+  cli();
+  //each sector is 512 bytes (or 256 words)
+  uint16_t *buf = (uint16_t *)op->buffer;
+  for(register size_t i=op->buffer_loc; i<op->buffer_loc+256;i++) {
+    if(i==op->buffer_loc+255) inb(ATA_STATUS(op->base_addr)); //read status byte to reset interrupt flag. We don't actually care about the values right now.
+    buf[i] = inw(ATA_DATA_REG(op->base_addr));
+  }
+
+  op->buffer_loc += 256;
+  ++op->sectors_read;
+  if(op->sectors_read>=op->sector_count) {
+    //reset the "pending operation" block for the next operation
+    //memset((void *)op, 0, sizeof(ATAPendingOperation));
+    op->type = ATA_OP_NONE;
+    //the operation is now completed
+    op->completed_func(ATA_STATUS_OK, op->buffer);
+  }
+
+  if(old_pd!=NULL) switch_paging_directory_if_required(old_pd);
+
+  sti();
+
+}
+
+/*
+This is an internal routine that writes a block of bytes to the disk from a buffer.
+It's called from `ata_pio_start_write` to write the first sector requested and then
+from `ata_complete_write_lowerhalf` for each subsequent write.
+Arguments: op - pointer to an ATAPendingOperation giving details of the operation in progress
+Returns: Nothing
+*/
+void ata_continue_write(ATAPendingOperation *op)
+{
+  vaddr old_pd = switch_paging_directory_if_required((vaddr)op->paging_directory);
+
+  //need to make sure interrupts are disabled, otherwise we trigger the next data packet
+  //before we stored the last word of this one, meaning that we miss data.
+  cli();
+
+  //kprintf("DEBUG sectors written %d sectors total to write %d\r\n", op->sectors_read, op->sector_count);
+
+  //each sector is 512 bytes (or 256 words)
+  uint16_t *buf = (uint16_t *)op->buffer;
+  for(register size_t i=op->buffer_loc; i<op->buffer_loc+256;i++) {
+    if(i==op->buffer_loc+255) inb(ATA_STATUS(op->base_addr)); //read status byte to reset interrupt flag. We don't actually care about the values right now.
+    outw(ATA_DATA_REG(op->base_addr), buf[i]);
+    io_wait();  //in ioports.asm
+  }
+
+  outb(ATA_COMMAND(op->base_addr), ATA_CMD_CACHE_FLUSH);  //make sure to flush the cache once the write has completed
+
+  op->buffer_loc += 256;
+  ++op->sectors_read;
+
+
+  if(old_pd!=NULL) switch_paging_directory_if_required(old_pd);
+
+  sti();
+}
+
+
+/*
+This is a "lower half", i.e. a routine that is run by the scheduler from the root
+kernel event loop. It's requested by the interrupt handler and does the actual work
+of taking bytes from memory and writing them to disk.
+Arguments: t - pointer to the SchedulerTask that was used to queue the event. The
+      "data" field can be assumed to be a pointer to ATAPendingOperation which
+      gives details of the event in progress
+Returns: Nothing
+*/
+void ata_complete_write_lowerhalf(SchedulerTask *t)
+{
+  ATAPendingOperation *op = (ATAPendingOperation *)t->data;
+
+  kprintf("DEBUG in disk write lower-half. Requested paging dir is 0x%x\r\n", op->paging_directory);
+
+  if(op->sectors_read>=op->sector_count) {
+    //the operation is now completed
+    op->completed_func(ATA_STATUS_OK, op->buffer);
+    //reset the "pending operation" block for the next operation
+    op->type = ATA_OP_NONE;
+  } else {
+      ata_continue_write(op);
+  }
 }
